Strict separation between client/session and session/doc messaging
This also improves the skeleton for where authentication/authorization checks will be made.  I decided to make session IDs be public since a new session is spawned for each connection to the WS url instead of ahead of time.  These changes show how the session layer is not just a pass-through.