Improved access convention and fixed bug
The default value for the permission struct is either: not valid because of user/circuit/link id being blank; no permission, because the access level is 0.  Hence, there is no reason to have a pointer return type to indicate the "not found" case.  I updated the AccessDriver to reflect this

I also fixed a bug in the model logic that would allow a low-permission user to demote a high-permission user.