Busing components now works
The only caveat is that the number of open/free InputPorts and OutputPorts for the components must be exact.

I.e. to bus 6 switches and a mux, the mux's OutputPort must have a wire connected.

There is a known bug where if you connect an encoder and decoder it connects all the ports due to ambiguity.